+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_service_specs/sidecar.py
@@ -11,7 +11,7 @@ from ....core.settings import AppSettings, DynamicSidecarSettings
 from ....models.schemas.constants import DYNAMIC_SIDECAR_SCHEDULER_DATA_LABEL
 from ....models.schemas.dynamic_services import SchedulerData, ServiceType
 from .._namespace import get_compose_namespace
-from ..volumes_resolver import DynamicSidecarVolumesPathsResolver
+from ..volumes import DynamicSidecarVolumesPathsResolver
 from .settings import update_service_params_from_settings
 
 log = logging.getLogger(__name__)
